modified layouts file structure
